+++ src/main/java/com/hhplus/concert/domain/payment/event/PaymentEvent.java
@@ -1,10 +1,11 @@
-package com.hhplus.concert.domain.payment;
+package com.hhplus.concert.domain.payment.event;
 
 import lombok.Getter;
+import lombok.Setter;
 
 public class PaymentEvent {
-
     @Getter
+    @Setter
     public static class Completed {
         private Long paymentId;
         private Long reservationId;
