+++ src/main/java/com/hhplus/concert/domain/payment/event/PaymentEventPublisher.java
@@ -1,4 +1,4 @@
-package com.hhplus.concert.domain.payment;
+package com.hhplus.concert.domain.payment.event;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
